Move subcommands to own folder
