+++ src/bin/why/commands/mod.rs
@@ -1,0 +1,3 @@
+mod build_executable;
+
+pub use build_executable::*;
